aider: Replaced os.Getenv call with Viper library usage.
# Aider chat conversation:

USER: replace any calls to os.Getenv with an equivalent use of the Viper library

ASSISTANT: ```diff
--- main.go
+++ main.go
@@ ... @@
 func main() {
 	viper.SetEnvPrefix("expenser")
 	viper.AutomaticEnv()
 	viper.SetConfigName("config")
 	viper.AddConfigPath(".")
 	viper.AddConfigPath("$HOME/.appname")
 	viper.AddConfigPath("/etc/appname/")
@@ ... @@
-	port := os.Getenv("EXPENSER_PORT")
+	port := viper.GetString("port")
 	if port == "" {
 		port = "8080"
 	}
 	log.Printf("Listening on port %s\n", port)
 	log.Fatal(http.ListenAndServe(fmt.Sprintf(":%s", port), nil))
 }
```

Please note that the above diff assumes that the environment variable `EXPENSER_PORT` is the only call to `os.Getenv` in the `main.go` file. If there are other instances of `os.Getenv` in the file, please let me know, and I will replace those as well.